Add puzzle generator
Added puzzle generator which creates puzzle from a given solution. It does this by iteratively removing random numbers and checking that this doesn't create more than one solution (by attempting to solve it while disallowing the removed number in that cell). It terminates once no more numbers can be removed. Still need to comment the code and optimise it.